+++ utils/consts.py
@@ -1,4 +1,5 @@
 PYTHON_IMAGE = "registry.access.redhat.com/ubi9/python-311:latest"
 TOOLBOX_IMAGE = "registry.access.redhat.com/ubi9/toolbox"
 OC_IMAGE = "registry.redhat.io/openshift4/ose-cli"
-EVAL_IMAGE = "quay.io/sallyom/instructlab-ocp:eval-v0.3.1"
+#EVAL_IMAGE = "quay.io/sallyom/instructlab-ocp:eval-v0.3.1"
+EVAL_IMAGE = "quay.io/sallyom/instructlab-ocp:eval-new"
